[#41] refactor: 약관동의 페이지 스타일 코드 분리
