+++ src/components/Auth/Register/RegisterDoc.js
@@ -1,90 +1,10 @@
 import React, { useState, useCallback } from 'react';
+import { useNavigate } from 'react-router';
 import styled from 'styled-components';
 import Button from 'components/Shared/Button';
 import * as S from 'components/Auth/styles';
-import { useNavigate } from 'react-router';
+import TitleSection from 'components/Shared/TitleSection';
 import { CREATE_ACCOUNT } from 'constant';
-
-const RegisterDocWrapper = styled.div`
-  @media screen and (max-width: ${(props) => props.theme.deviceSizes.mobileL}) {
-    background-color: ${(props) => props.theme.colors.white};
-    width: inherit;
-    top: 0;
-    position: absolute;
-  }
-`;
-
-const BackButton = styled.button`
-  display: none;
-
-  @media screen and (max-width: ${(props) => props.theme.deviceSizes.mobileL}) {
-    display: block;
-    width: 24px;
-    height: 24px;
-    background: url(/asset/chevron-left.svg);
-    background-size: 24px;
-    left: 8px;
-    position: absolute;
-  }
-`;
-
-const RegisterDocSection = styled.div`
-  @media screen and (max-width: ${(props) => props.theme.deviceSizes.mobileL}) {
-    top: 62px;
-    position: relative;
-  }
-`;
-
-const RegisterDocDescSection = styled.section`
-  display: none;
-
-  @media screen and (max-width: ${(props) => props.theme.deviceSizes.mobileL}) {
-    display: block;
-    margin-top: 24px;
-    margin-bottom: 9px;
-    position: relative;
-  }
-`;
-
-const Title = styled.p`
-  margin-bottom: 8px;
-  color: ${(props) => props.theme.colors.darkgray};
-  font-size: 16px;
-  font-weight: 500;
-  line-height: normal;
-`;
-
-const Desc = styled.p`
-  font-size: 12px;
-  font-weight: normal;
-  line-height: normal;
-  color: ${(props) => props.theme.colors.gray}; ;
-`;
-
-const Bottom = styled.div`
-  position: fixed;
-  bottom: 40px;
-`;
-
-const ProcessBarSection = styled.div`
-  display: flex;
-  padding: 16px 0;
-  justify-content: center;
-`;
-
-const ProgressCircle = styled.div`
-  background-color: ${({ isOnProgress, ...props }) =>
-    isOnProgress ? props.theme.colors.yellow : props.theme.colors.silver};
-  width: 8px;
-  height: 8px;
-  flex-grow: 0;
-  margin: 0 16px 0 0;
-  border-radius: 100%;
-
-  :last-child {
-    margin: 0;
-  }
-`;
 
 const NextButton = styled(Button)`
   margin-top: 0;
@@ -121,22 +41,15 @@ const RegisterDoc = ({ checked }) => {
     navigate(CREATE_ACCOUNT);
   };
 
-  const shiftBack = () => {
-    navigate(-1);
-  };
-
   return (
-    <RegisterDocWrapper>
-      <S.Title>
-        <BackButton onClick={shiftBack} />
-        회원가입
-      </S.Title>
+    <S.ContentWrapper>
+      <TitleSection />
 
-      <RegisterDocSection>
-        <RegisterDocDescSection>
-          <Title>약관 동의</Title>
-          <Desc>원활한 사용을 위하여 필수 약관 동의가 필요합니다.</Desc>
-        </RegisterDocDescSection>
+      <S.ContentSection>
+        <S.ContentDescSection>
+          <S.DescTitle>약관 동의</S.DescTitle>
+          <S.DescText>원활한 사용을 위하여 필수 약관 동의가 필요합니다.</S.DescText>
+        </S.ContentDescSection>
 
         <S.AllAgree>
           <S.Agree>
@@ -177,18 +90,18 @@ const RegisterDoc = ({ checked }) => {
           합니다
         </S.AuthDoc>
 
-        <Bottom>
-          <ProcessBarSection>
-            <ProgressCircle isOnProgress={true} />
-            <ProgressCircle />
-          </ProcessBarSection>
+        <S.BottomProgressBar>
+          <S.ProgressBarSection>
+            <S.ProgressCircle isOnProgress={true} />
+            <S.ProgressCircle />
+          </S.ProgressBarSection>
 
           <NextButton onClick={onClick} disabled={checkedList.length !== dataLists.length} type="button">
             다음
           </NextButton>
-        </Bottom>
-      </RegisterDocSection>
-    </RegisterDocWrapper>
+        </S.BottomProgressBar>
+      </S.ContentSection>
+    </S.ContentWrapper>
   );
 };
